+++ CHANGELOG.md
@@ -3,7 +3,13 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------
+##### Fixed
+- remove existing content form external instances
+
 [3.5.2] - 2015-04-08
+---------------
 ##### Added
 - support for current()